requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymongo==4.6.0
google-generativeai==0.3.1
python-dotenv==1.0.0
vaderSentiment==3.3.2
bson==0.5.10 